backend/server/requirements.txt
django==4.2.7
djangorestframework==3.14.0
markdown==3.5.1
django-filter==23.3
djoser==2.2.0
django_cors_headers==4.3.0
pillow==10.1.0